Payment status fix, price double to  BigDecimal

+++ src/net/epam/study/service/ChangeOrderService.java
@@ -1,8 +1,10 @@
 package net.epam.study.service;
 
+import java.math.BigDecimal;
+
 public interface ChangeOrderService {
     void deleteOrderItem(String item, String login) throws ServiceException;
     void addToOrder(String name, String price, String time);
-    double getTotal(String login) throws ServiceException;
+    BigDecimal getTotal(String login) throws ServiceException;
     StringBuilder getOrder();
 }
